split the components in Tag.js into different files and cleanup
